Pad the begining and end of a vector. If one wants to evaluate a text string, use `<< >>` to quote the code to be evaluted.

Pad the begining and end of a vector. If one wants to evaluate a text string, use `<< >>` to quote the code to be evaluted.

Usage

padding(txt, pattern)

Arguments

txt

string vector with length of one.

pattern

pattern to pad.

Value

padded string